Refactor to join A and D tests.
